fix: carousel 수정
